[Evaluation] fixes initialization
