+++ src/main/evaluation/Resources/config/services/messenger.yml
@@ -1,4 +1,10 @@
 services:
+    Claroline\EvaluationBundle\Messenger\InitializeResourceEvaluationsHandler:
+        tags: [ messenger.message_handler ]
+        arguments:
+            - '@Claroline\AppBundle\Persistence\ObjectManager'
+            - '@Claroline\EvaluationBundle\Manager\ResourceEvaluationManager'
+
     Claroline\EvaluationBundle\Messenger\InitializeSequenceEvaluationsHandler:
         tags: [ messenger.message_handler ]
         arguments:
@@ -12,12 +18,6 @@ services:
             - '@Claroline\AppBundle\Persistence\ObjectManager'
             - '@Claroline\EvaluationBundle\Manager\WorkspaceEvaluationManager'
             - '@Claroline\EvaluationBundle\Manager\SequenceEvaluationManager'
-
-    Claroline\EvaluationBundle\Messenger\UpdateResourceEvaluationsHandler:
-        tags: [messenger.message_handler]
-        arguments:
-            - '@Claroline\AppBundle\Persistence\ObjectManager'
-            - '@Claroline\EvaluationBundle\Manager\ResourceEvaluationManager'
 
     Claroline\EvaluationBundle\Messenger\RecomputeResourceEvaluationsHandler:
         tags: [ messenger.message_handler ]
